backend/requirements.txt
# Core
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
python-multipart>=0.0.9

# Database
sqlalchemy>=2.0.0
alembic>=1.13.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
pgvector>=0.3.6

# Redis / Celery
redis>=5.0.0
celery[redis]>=5.4.0

# Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
google-auth>=2.30.0
google-auth-oauthlib>=1.2.0

# HTTP
httpx>=0.27.0

# AI / ML
anthropic>=0.30.0
openai>=1.58.0

# Pipeline
python-pptx>=1.0.2
Pillow>=11.0.0
deepl>=1.18.0
google-cloud-translate>=3.18.0
google-cloud-texttospeech>=2.21.0

# Storage
boto3>=1.34.0
